Message Mutation feature
Added Message Mutation option. So by sending message from the app a data from the up is uploading to the server. At this point the app can be use for communication but still a lot of to do. There is still problem with refreshing data form server...so pulling not working well also I need to implement some user feature because all messages is shown as Syrius (me??)